+++ openbb_platform/extensions/equity/openbb_equity/equity_views.py
@@ -2,9 +2,9 @@
 
 from typing import Any, Dict, Tuple
 
+from openbb_charting.charts.price_historical import price_historical
+from openbb_charting.charts.price_performance import price_performance
 from openbb_charting.core.openbb_figure import OpenBBFigure
-from openbb_charting.utils.price_historical import price_historical
-from openbb_charting.utils.price_performance import price_performance
 
 
 class EquityViews:
