+++ openbb_platform/extensions/etf/openbb_etf/etf_views.py
@@ -3,10 +3,10 @@
 from typing import Any, Dict, Tuple, Union
 
 import pandas as pd
+from openbb_charting.charts.generic_charts import bar_chart
+from openbb_charting.charts.price_historical import price_historical
+from openbb_charting.charts.price_performance import price_performance
 from openbb_charting.core.openbb_figure import OpenBBFigure
-from openbb_charting.utils.generic_charts import bar_chart
-from openbb_charting.utils.price_historical import price_historical
-from openbb_charting.utils.price_performance import price_performance
 from openbb_core.app.utils import basemodel_to_df
 from plotly.graph_objs import Figure
