+++ openbb_platform/extensions/technical/openbb_technical/technical_views.py
@@ -3,12 +3,12 @@
 from typing import Any, Dict, Tuple
 
 import pandas as pd
+from openbb_charting.charts import relative_rotation
 from openbb_charting.core.chart_style import ChartStyle
 from openbb_charting.core.openbb_figure import OpenBBFigure
 from openbb_charting.core.plotly_ta.ta_class import PlotlyTA
 from openbb_charting.core.to_chart import to_chart
 from openbb_charting.styles.colors import LARGE_CYCLER
-from openbb_charting.utils import relative_rotation
 from openbb_core.app.utils import basemodel_to_df
