+++ openbb_platform/obbject_extensions/charting/openbb_charting/charting.py
@@ -25,11 +25,11 @@ from openbb_core.app.utils import basemodel_to_df, convert_to_basemodel
 from openbb_core.provider.abstract.data import Data
 from plotly.graph_objs import Figure
 
+from openbb_charting.charts.generic_charts import bar_chart, line_chart
+from openbb_charting.charts.helpers import get_charting_functions
 from openbb_charting.core.backend import Backend, create_backend, get_backend
 from openbb_charting.core.openbb_figure import OpenBBFigure
 from openbb_charting.query_params import ChartParams, IndicatorsParams
-from openbb_charting.utils.generic_charts import bar_chart, line_chart
-from openbb_charting.utils.helpers import get_charting_functions
 
 
 class Charting:
